+++ axum-extra/CHANGELOG.md
@@ -15,14 +15,18 @@ and this project adheres to [Semantic Versioning].
 - **changed:** For methods that accept some `S: Service`, the bounds have been
   relaxed so the response type must implement `IntoResponse` rather than being a
   literal `Response`
+- **added:** Support chaining handlers with `HandlerCallWithExtractors::or` ([#1170])
+- **change:** axum-extra's MSRV is now 1.60 ([#1239])
 
 [#1086]: https://github.com/tokio-rs/axum/pull/1086
 [#1119]: https://github.com/tokio-rs/axum/pull/1119
+[#1170]: https://github.com/tokio-rs/axum/pull/1170
+[#1239]: https://github.com/tokio-rs/axum/pull/1239
 
 # 0.3.5 (27. June, 2022)
 
 - **added:** Add `JsonLines` for streaming newline delimited JSON ([#1093])
-- **change:** axum's MSRV is now 1.56 ([#1098])
+- **change:** axum-extra's MSRV is now 1.56 ([#1098])
 
 [#1093]: https://github.com/tokio-rs/axum/pull/1093
 [#1098]: https://github.com/tokio-rs/axum/pull/1098